整理代码

+++ Assets/Scripts/Frame/CommandSystem/IDelayCmdWatcher.cs
@@ -5,5 +5,5 @@ public interface IDelayCmdWatcher
 	void addDelayCmd(Command cmd);
 	void onCmdStarted(Command cmd);
 	void interruptAllCommand();
-	void interruptCommand(int assignID, bool showError);
+	void interruptCommand(ulong assignID, bool showError);
 }
